requirements.txt
fastapi
aiogram==3.*
sqlalchemy[asyncio]
asyncpg
apscheduler
pydantic
uvicorn
alembic
psycopg2-binary
httpx
python-jose
passlib
python-multipart
aiogram
python-dotenv
pytest
pytest-asyncio